refactor: move node dependencies from models to helpers
