+++ packages/insomnia-scripting-environment/src/objects/send-request.ts
@@ -1,5 +1,5 @@
 import type { CurlRequestOutput } from 'insomnia/src/main/network/libcurl-promise';
-import { readCurlResponse } from 'insomnia/src/models/response';
+import { readCurlResponse } from 'insomnia/src/models/helpers/response-operations';
 import type { Settings } from 'insomnia/src/models/settings';
 import { Cookie } from 'tough-cookie';
 import { v4 as uuidv4 } from 'uuid';
