+++ packages/insomnia/src/ui/components/mcp/mcp-pane.tsx
@@ -681,5 +681,3 @@ const CollectionGridListItem = (props: {
     </GridListItem>
   );
 };
-
-export default McpPane;
